An interactive bounding-box tool for Mastodon.
Installs an interactive bounding-box tool on a BDV.

The feature consists of a bounding-box model, a dialog and an overlay added to
the BDV. The user can edit the bounding-box using a GUI panel, or directly
interacting with the overlay using an 'edit' mode. The user switches between
normal ('visualization') and 'edit' modes by pressing the ESC key. In 'edit'
mode, the mouse button is used to drag the corners of the bounding-box.

Adapted from code by Tobias Pietzsch and others taken in the BDV core packages.