Semi-auto tracker can specify the resolution level it operates on.
Using the level 0 (highest res, lower level index) on large image
shows that we are a bit sub-optimal in memory management. Even if
we perform the detection on very small portion of the image (lines
195-260), we still load the whole image for the current time-point
(line 187), which takes a long time and a lot of RAM for large images.

A good strategy would be to only load the chunks of data around the
ROI. This good be exploited as well in TrackMate detectors.