Do not load the image completely if we can use a Volatile image.
With the flag ImgLoaderHints.LOAD_COMPLETELY, we were loading the
image completely regardless of whether we had a BDV HDF5 image
stored on the HD. Not using this flag then lets the Hdf5ImageLoader$
SetupImgLoader creates a cached image that loads the blocks on demand.

In practice, this turns in having the semi-automatic tracker perform
very quickly even on large images at full resolution, since we only
query a small ROI of said image at each step.

Since this correction is common to also detectors, we expect them to
benefit from it as well, provided they use small ROIs.