Rework entirely the particle-linking framework.
At the general level, separate linkers that can operate sequentially
and those who need a graph to operate. The input changes, for one,
the LinkCreator is enough (no need to look on the data), for the other
we need to specify a graph (on which we can query track segments, etc...).

Reflect these on the packages.

On the Mamut level, kill generality with a ModelGraph and a specific
LinkCreator.