The detector preview uses its own model and view.
We create a new BDV window for preview, with a local model, that is
not registered in the WindowManager. Modications done during the
preview then do not affect the app model. As a price to pay: the preview
bdv window is stripped down of the niceties that are found for 1st class
bdv views.