Semi-auto tracker relies on the smallest spot source radius.
Before this commit, we used the largest radius of the ellipsoid.
For classical microscopy images, the spot can be very elongated in
one direction, which will lead in having very large radius, therefore
very large smoothing values for the DoG detection. It is best to use
the smallest radius instead.