Change how we decide at what resolution level to perform detection.
Previously, it worked as follow:
Iterate from the lowest level of the pyramid (highest resolution)
to the top (most downsampled). When the pixel size of an object in
any dimension goes below the limit, stop at this level.

This is not optimal for microscopy data. Typically, even with LSFMs,
the Z sampling can be much lower than in X and Y. The pixel size in Z
is them much larger than in X and Y. For instance on a 25x water
objective imaging on a 2048x2048 sCMOS camera the pixel size in X, Y
and Z are respectively [ 0.35, 0.35, 1.5 ] µm. This is going to be a
common case for microscopists using modern cameras.

There is a factor 4 between X and Z pixel sizes. The BDV conversion tool
picks this up correctly, and proposes the following mipmap scales:
0: [ 1 1 1
1:   2 2 1
2:   4 4 1
3:   8 8 2 ]

If we are to detect nuclei that are about 3µm in radius, we would like
them to have be at most 2.5 pixels in all directions at the optimal
resolution level.

But with the previous level determination algorithm things happened like
this: "
- Iterate to level 0.
- At this level, the size of my object is [ 9.6, 9.6, 2.0 ] pixels.
- The Z dimension has a size 2.0 pixels, below 2.5 pixels, so we stop
here and return level 0.
"

But if we would have gone to level 1, the object size would have been
the same size in Z, but with a lower size in X and Y, permitting
downsampling.

This new algorithm deals with this by doing the following: "
- Iterate to level i.
- Compute the size of object in all dimensions.
- Iterate to dimension d.
- If the size of object at this dimension is smaller than the limit, then
we stop at this level, but only if:
	- we lower the size of the object in this dimension even more
	(compared with previous level).
	- all dimensions are below the limit for the first time.
"

With the previous example, the algorithm performs as follow: "
- Iterate to level 0.
- At this level, the size of my object is [ 9.6, 9.6, 2.0 ] pixels.
- The Z dimension has a size 2.0 pixels, below 2.5 pixels. But:
	- this is the first time,
	- and the other dimensions are above the limit.
- Iterate to level 1.
- At this level, the size of my object is [ 4.8, 4.8, 2.0 ] pixels.
- The Z dimension has a size 2.0 pixels, below 2.5 pixels. But:
	- this is NOT the first time, but we did not decrease
	its size more.
	- and the other dimensions are still above the limit.
- Iterate to level 2.
- At this level, the size of my object is [ 2.4, 2.4, 2.0 ] pixels.
- All dimensions are below the limit -> we stop there.
"

There is probably a more compact way to implement this algorithm,
but this one expresses what we have in mind.