Hopefully fix physical scaling.
Added a function getPhysicalSizeOfGlobalUnit() that determines the size of
a unit in the global coordinate system. This works by taking the voxel
dimension of a given source (which specifies how big a voxel at level 0
of this source is in physical units). Then the transformation of level 0
of the source into the global coordinate system is applied to get the
global unit length. Technically, the voxel dimension and source transform
specified in the BDV XML may be inconsistent (for example, if the transform
contains shearing, or would make the global system non-isotropic). Not to
metion that different sources may specify mutually inconsisten information.
Therefore, some compromise needs to be made, and this compromise is (here
and in other places) to do the calculation with the pixel width (dimension
0) from the voxel dimensions of a specified setup.

This physicalSizeOfGlobalUnit needs to be considered when computing the
physical calibration of a particular resolution level of a particular
setup. It also needs to be taken into account when setting the radius of
a detected spot (which is specified in global units).