Add a separator in Tracking menu.
To separate semi-auto tracker from the wizard-based plugin.
This works (the separator is at the right place despite the menu
items being specified in different classes) because plugins are
discovered and added in a deterministic manner. Indeed, the plugins
after discovery are returned in a list sorted by class name.