Change how default parameters are specified.
Before this commit, detection and tracking parameters were specified
via the UI descriptor, which made the methods not available to the
scripting interface. They were hardcoded.

Making detectors and linkers extensible require having a means to
specify the default parameters somewhere conveninent, and now this
is done in the SpotDetectorOp and SpotLinkerOp interfaces.
This requires some gymnastics in the UI descriptor, but that's
life.